Refatoração models
