--- PyNFSe/utils/models/__init__.py
@@ -1,7 +1,0 @@
-from .lote_rps import LoteRPS
-from .servico import Servico
-from .tomador import Tomador
-from .prestador import Prestador
-from .rps import RPS
-from .pedido_cancelamento import PedidoCancelamentoNFSe
-from .configuracao import Ambiente, Configuracao
