--- PyNFSe/utils/models/base.py
@@ -1,7 +1,0 @@
-
-
-class Entidade(object):
-
-    def __init__(self, **kwargs):
-        for key, value in kwargs.items():
-            setattr(self, key, value)
