--- PyNFSe/utils/models/lote_rps.py
@@ -1,9 +1,0 @@
-from simple_model import Model
-
-
-class LoteRPS(Model):
-    identificador: str
-    numero_lote: int
-    cnpj: str
-    inscricao_municipal: str
-    lista_rps: list
