--- PyNFSe/utils/models/pedido_cancelamento.py
@@ -1,10 +1,0 @@
-from .base import Entidade
-
-
-class PedidoCancelamentoNFSe(Entidade):
-
-    identificador = str()
-    prestador = None
-    numero_nota = int()
-    codigo_municipio = str()
-    codigo_cancelamento = str()
