--- PyNFSe/utils/models/prestador.py
@@ -1,7 +1,0 @@
-from simple_model import Model
-
-
-class Prestador(Model):
-
-    cnpj: str
-    inscricao_municipal: str
