--- PyNFSe/utils/models/servico.py
@@ -1,56 +1,0 @@
-import typing
-
-from simple_model import Model
-
-from decimal import Decimal
-
-
-class Servico(Model):
-
-    valor_servico: Decimal = Decimal()
-    iss_retido: int  # 1 - Sim, 2 - Não
-    item_lista: str
-    discriminacao: str
-    codigo_municipio: str
-    codigo_cnae: typing.Any = None
-    codigo_tributacao_municipio: typing.Any = None
-
-    valor_deducoes: Decimal = Decimal()
-    valor_pis: Decimal = Decimal()
-    valor_cofins: Decimal = Decimal()
-    valor_inss: Decimal = Decimal()
-    valor_ir: Decimal = Decimal()
-    valor_csll: Decimal = Decimal()
-    outras_retencoes: Decimal = Decimal()
-    aliquota: Decimal = Decimal()
-    desconto_incondicionado: Decimal = Decimal()
-    desconto_condicionado: Decimal = Decimal()
-
-    def __init__(self, **kwargs):
-
-        super(Servico, self).__init__(**kwargs)
-
-        self.base_calculo = self._calcular_base_calculo()
-        self.valor_iss = self._calcular_iss()
-        self.valor_iss_retido = Decimal() if self.iss_retido == 2 else self.valor_iss
-        self.valor_liquido = self._calcular_valor_liquido()
-
-    def _calcular_base_calculo(self):
-
-        base_calculo = self.valor_servico - self.valor_deducoes - self.desconto_incondicionado
-
-        return base_calculo
-
-    def _calcular_valor_liquido(self):
-
-        valor_liquido = self.valor_servico - self.valor_pis - self.valor_cofins - self.valor_inss -\
-                        self.valor_ir - self.valor_csll - self.outras_retencoes - self.valor_iss_retido -\
-                        self.desconto_incondicionado - self.desconto_condicionado
-
-        return Decimal(valor_liquido)
-
-    def _calcular_iss(self):
-
-        valor_iss = self.base_calculo * self.aliquota
-
-        return Decimal(valor_iss).quantize(Decimal(10) ** -2)
